LPD-58182 Inline WildcardQueryTranslator/WildcardQueryTranslatorImpl
