--- modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/query/WildcardQueryTranslatorImpl.java
@@ -1,34 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.elasticsearch7.internal.query;
-
-import com.liferay.portal.search.query.WildcardQuery;
-
-import org.elasticsearch.index.query.QueryBuilder;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.WildcardQueryBuilder;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Miguel Angelo Caldas Gallindo
- */
-@Component(service = WildcardQueryTranslator.class)
-public class WildcardQueryTranslatorImpl implements WildcardQueryTranslator {
-
-	@Override
-	public QueryBuilder translate(WildcardQuery wildcardQuery) {
-		WildcardQueryBuilder wildcardQueryBuilder = QueryBuilders.wildcardQuery(
-			wildcardQuery.getField(), wildcardQuery.getValue());
-
-		if (wildcardQuery.getRewrite() != null) {
-			wildcardQueryBuilder.rewrite(wildcardQuery.getRewrite());
-		}
-
-		return wildcardQueryBuilder;
-	}
-
-}
