+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/query/BaseQueryVisitor.java
@@ -5,7 +5,10 @@
 
 package com.liferay.portal.search.solr8.internal.query;
 
+import com.liferay.petra.string.CharPool;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.search.BooleanQuery;
+import com.liferay.portal.kernel.search.QueryTerm;
 import com.liferay.portal.kernel.search.TermQuery;
 import com.liferay.portal.kernel.search.TermRangeQuery;
 import com.liferay.portal.kernel.search.WildcardQuery;
@@ -18,7 +21,11 @@ import com.liferay.portal.kernel.search.generic.MultiMatchQuery;
 import com.liferay.portal.kernel.search.generic.NestedQuery;
 import com.liferay.portal.kernel.search.generic.StringQuery;
 import com.liferay.portal.kernel.search.query.QueryVisitor;
+import com.liferay.portal.kernel.util.StringUtil;
 
+import org.apache.lucene.index.Term;
+import org.apache.lucene.queryparser.classic.QueryParser;
+import org.apache.lucene.search.BoostQuery;
 import org.apache.lucene.search.Query;
 
 import org.osgi.service.component.annotations.Reference;
@@ -85,7 +92,18 @@ public abstract class BaseQueryVisitor implements QueryVisitor<Query> {
 
 	@Override
 	public Query visitQuery(WildcardQuery wildcardQuery) {
-		return wildcardQueryTranslator.translate(wildcardQuery);
+		QueryTerm queryTerm = wildcardQuery.getQueryTerm();
+
+		Query query = new org.apache.lucene.search.WildcardQuery(
+			new Term(
+				_escapeSpaces(queryTerm.getField()),
+				_escape(queryTerm.getValue())));
+
+		if (!wildcardQuery.isDefaultBoost()) {
+			return new BoostQuery(query, wildcardQuery.getBoost());
+		}
+
+		return query;
 	}
 
 	@Reference
@@ -121,7 +139,42 @@ public abstract class BaseQueryVisitor implements QueryVisitor<Query> {
 	@Reference
 	protected TermRangeQueryTranslator termRangeQueryTranslator;
 
-	@Reference
-	protected WildcardQueryTranslator wildcardQueryTranslator;
+	private String _escape(String value) {
+		int x = 0;
+		int y = 0;
+
+		int length = value.length();
+
+		StringBuilder sb = new StringBuilder(length * 2);
+
+		while (y < length) {
+			char c = value.charAt(y);
+
+			if ((c == CharPool.QUESTION) || (c == CharPool.SPACE) ||
+				(c == CharPool.STAR)) {
+
+				sb.append(QueryParser.escape(value.substring(x, y)));
+
+				if (c == CharPool.SPACE) {
+					sb.append(CharPool.BACK_SLASH);
+				}
+
+				sb.append(c);
+
+				x = y + 1;
+			}
+
+			y++;
+		}
+
+		sb.append(QueryParser.escape(value.substring(x)));
+
+		return sb.toString();
+	}
+
+	private String _escapeSpaces(String value) {
+		return StringUtil.replace(
+			value, CharPool.SPACE, StringPool.BACK_SLASH + StringPool.SPACE);
+	}
 
 }
