--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/query/WildcardQueryTranslator.java
@@ -1,20 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.query;
-
-import com.liferay.portal.kernel.search.WildcardQuery;
-
-import org.apache.lucene.search.Query;
-
-/**
- * @author André de Oliveira
- * @author Miguel Angelo Caldas Gallindo
- */
-public interface WildcardQueryTranslator {
-
-	public Query translate(WildcardQuery wildcardQuery);
-
-}
