--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/query/WildcardQueryTranslatorImpl.java
@@ -1,82 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.query;
-
-import com.liferay.petra.string.CharPool;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.search.QueryTerm;
-import com.liferay.portal.kernel.search.WildcardQuery;
-import com.liferay.portal.kernel.util.StringUtil;
-
-import org.apache.lucene.index.Term;
-import org.apache.lucene.queryparser.classic.QueryParser;
-import org.apache.lucene.search.BoostQuery;
-import org.apache.lucene.search.Query;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author André de Oliveira
- * @author Miguel Angelo Caldas Gallindo
- */
-@Component(service = WildcardQueryTranslator.class)
-public class WildcardQueryTranslatorImpl implements WildcardQueryTranslator {
-
-	@Override
-	public Query translate(WildcardQuery wildcardQuery) {
-		QueryTerm queryTerm = wildcardQuery.getQueryTerm();
-
-		Query query = new org.apache.lucene.search.WildcardQuery(
-			new Term(
-				_escapeSpaces(queryTerm.getField()),
-				_escape(queryTerm.getValue())));
-
-		if (!wildcardQuery.isDefaultBoost()) {
-			return new BoostQuery(query, wildcardQuery.getBoost());
-		}
-
-		return query;
-	}
-
-	private String _escape(String value) {
-		int x = 0;
-		int y = 0;
-
-		int length = value.length();
-
-		StringBuilder sb = new StringBuilder(length * 2);
-
-		while (y < length) {
-			char c = value.charAt(y);
-
-			if ((c == CharPool.QUESTION) || (c == CharPool.SPACE) ||
-				(c == CharPool.STAR)) {
-
-				sb.append(QueryParser.escape(value.substring(x, y)));
-
-				if (c == CharPool.SPACE) {
-					sb.append(CharPool.BACK_SLASH);
-				}
-
-				sb.append(c);
-
-				x = y + 1;
-			}
-
-			y++;
-		}
-
-		sb.append(QueryParser.escape(value.substring(x)));
-
-		return sb.toString();
-	}
-
-	private String _escapeSpaces(String value) {
-		return StringUtil.replace(
-			value, CharPool.SPACE, StringPool.BACK_SLASH + StringPool.SPACE);
-	}
-
-}
